# backend/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.BulkFailure:
    properties:
      error:
        type: string
      id:
        type: string
    type: object
  handlers.BulkMetadataRequest:
    properties:
      ids:
        items:
          type: string
        type: array
      source:
        description: |-
          Source optionally pins the fetch to one named metadata source
          (comicinfo, filename, e-hentai, nhentai). Omit it to run the normal
          priority-ordered pipeline.
        type: string
    type: object
  handlers.BulkProgressRequest:
    properties:
      ids:
        items:
          type: string
        type: array
      read:
        description: |-
          Read true marks every archive finished (its own last page); false
          clears progress entirely, the same as DELETE .../{id}/progress.
        type: boolean
    type: object
  handlers.BulkResult:
    properties:
      failed:
        items:
          $ref: '#/definitions/handlers.BulkFailure'
        type: array
      requested:
        type: integer
      succeeded:
        type: integer
    type: object
  handlers.BulkUpdateArchivesRequest:
    properties:
      artists:
        items:
          type: string
        type: array
      category:
        type: string
      characters:
        items:
          type: string
        type: array
      circles:
        items:
          type: string
        type: array
      ids:
        items:
          type: string
        type: array
      language:
        type: string
      parodies:
        items:
          type: string
        type: array
      tags:
        items:
          type: string
        type: array
    type: object
...
  response.Error:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Self-hosted doujinshi library manager API
  title: Shoka API
  version: "1.0"
paths:
...
  /api/archives/bulk:
    patch:
      consumes:
      - application/json
      description: Relations (artists/tags/parodies/circles/characters) MERGE into
        what each archive already has rather than replacing it, so an archive keeps
        its own tags and gains the submitted ones; de-duplication is case-insensitive
        and submitted names are snapped to the spelling already in use. Because it
        only adds, an empty array is a no-op rather than "clear" and bulk removal
        is unsupported. Scalars (language/category) are SET, overwriting the current
        value; omit or null them to leave them alone, send "" to blank them. Title,
        summary and release_date are intentionally not editable here since they describe
        one specific archive - use PATCH /api/archives/{id} for those. Each archive
        is applied in its own transaction, so a failure part-way through leaves earlier
        archives updated and reports the rest in "failed".
      parameters:
      - description: Archives and relations to add
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.BulkUpdateArchivesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.BulkResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      summary: Update shared metadata on archives in bulk
      tags:
      - archives
  /api/archives/bulk/metadata:
    post:
      consumes:
      - application/json
      description: 'Enqueues the same metadata job the scanner uses for newly added
        archives: local sources run first and are applied, then a remote job is chained
        to fill remaining gaps. Pass "source" to pin the fetch to a single named source
        instead, in which case only that source runs and nothing is chained; network-backed
        sources are queued on the low-concurrency worker so a large batch doesn''t
        hammer them. A source that is disabled for a given archive''s library is skipped
        rather than retried. Returns as soon as the jobs are queued - "succeeded"
        counts archives queued, not archives whose metadata has been written yet.
        Unlike the single-archive fetch endpoints there is no preview/confirm step,
        so results are applied automatically.'
      parameters:
      - description: Archives to identify, optionally pinned to one source
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.BulkMetadataRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/handlers.BulkResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      summary: Fetch and apply metadata for archives in bulk
      tags:
      - archives
  /api/archives/bulk/progress:
    put:
      consumes:
      - application/json
      description: Marking read jumps each archive to its own last page and flags
        it completed; marking unread deletes the progress row outright (page 0 would
        still count as in-progress). Ids that don't resolve to an archive are reported
        in "failed" rather than failing the request.
      parameters:
      - description: Archives and desired read state
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.BulkProgressRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.BulkResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      summary: Mark archives read or unread in bulk
      tags:
      - archives
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"